move rootModel files to session dir, rename as session files
