+++ packages/jbrowse-web/src/JBrowse.js
@@ -10,7 +10,7 @@ import '@gmod/jbrowse-core/fonts/material-icons.css'
 import PluginManager from '@gmod/jbrowse-core/PluginManager'
 import { openLocation } from '@gmod/jbrowse-core/util/io'
 import corePlugins from './corePlugins'
-import RootModelFactory from './rootModel'
+import RootModelFactory from './session/sessionModelFactory'
 import App from './ui/App'
 import Theme from './ui/theme'
