+++ packages/jbrowse-web/src/session/AssemblyManager.js
