+++ packages/jbrowse-web/src/session/assemblyConfigSchemas.js
