+++ packages/jbrowse-web/src/session/sessionConfigFactory.js
