+++ packages/jbrowse-web/src/session/sessionModel.test.js
@@ -1,5 +1,5 @@
 import { getSnapshot } from 'mobx-state-tree'
-import { createTestEnv } from './JBrowse'
+import { createTestEnv } from '../JBrowse'
 
 jest.mock('shortid', () => ({ generate: () => 'testid' }))
 
@@ -32,7 +32,7 @@ test('can load configuration from a config object', async () => {
 
 test('can load configuration from a file', async () => {
   const { rootModel } = await createTestEnv({
-    localPath: require.resolve('../test_data/config_volvox_mainthread.json'),
+    localPath: require.resolve('../../test_data/config_volvox_mainthread.json'),
   })
 
   expect(getSnapshot(rootModel.configuration)).toMatchSnapshot()
